+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Ansible-Alicloud Change Log
 ## 1.10.0 (Unreleased)
 
+IMPROVEMENTS:
+
+- Add creation_time attribute in ali_image_facts results ([#164](https://github.com/alibaba/ansible-provider/pull/164))
+
 BUG FIXES:
 
+- Fix grouping dynamic inventory on security group bug ([#165](https://github.com/alibaba/ansible-provider/pull/165))
 - Default to describe all instances in all regions ([#158](https://github.com/alibaba/ansible-provider/pull/158))
 - Fix ecs_instance_filters last page describe bug ([#157](https://github.com/alibaba/ansible-provider/pull/157))
 - Fix ecs_instance_filters's page_size bug ([#156](https://github.com/alibaba/ansible-provider/pull/156))